% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Animals}
\alias{Animals}
\title{Brain and Body Weights for 28 Species (from package MASS)}
\format{
\describe{
    \item{\code{body}}{
      body weight in kg.
    }
    \item{\code{brain}}{
      brain weight in g.
    }
  }
}
\source{
P. J. Rousseeuw  and A. M. Leroy (1987)
  \emph{Robust Regression and Outlier Detection.}
  Wiley, p. 57.
}
\usage{
Animals
}
\description{
Average brain and body weights for 28 species of land animals
}
\note{
The name \code{Animals} avoids conflicts with a system dataset \code{animals} in S-PLUS 4.5 and later.
}
\references{
Venables, W. N. and Ripley, B. D. (1999)
 \emph{Modern Applied Statistics with S-PLUS.} Third
 Edition. Springer.
}
\keyword{datasets}
