% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_param.R
\name{convert_betaXA_to_betaXB}
\alias{convert_betaXA_to_betaXB}
\title{Convert parameters between equivalent design matrices}
\usage{
convert_betaXA_to_betaXB(XA, XB, betaXA)
}
\arguments{
\item{XA}{The orignal design matrix}

\item{XB}{A design matrix equivalent to the original design matrix but with different parameterization}

\item{betaXA}{The vector of coefficients for the original design matrix}
}
\value{
The vector of coefficients for the design matrix XB
}
\description{
This function allows for the conversion of parameters between design matrices
that are equivalent (same data, same rank) but that have been parameterized
differently. This is usefull to find conversion of parameters for different
contrasts matrices (e.g. contrast treament vs contrast sum), including
conversion between \code{poly(..., raw = FALSE)} and
\code{poly(..., raw = TRUE)}.
}
\examples{
xa <- model.matrix(object = ~ x, data = data.frame(x = factor(c("A", "B", "C"))),
                   contrasts.arg = list(x = "contr.treatment"))
xb <- model.matrix(object = ~ x, data = data.frame(x = factor(c("A", "B", "C"))),
                   contrasts.arg = list(x = "contr.sum"))
beta_xa <- c(0.5, 2, 1)
beta_xb <- convert_betaXA_to_betaXB(XA = xa, XB = xb, betaXA = beta_xa)
## test that we get same predictions as we should:
xa \%*\% beta_xa
xb \%*\% beta_xb

}
