% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\name{get_vignettes}
\alias{get_vignettes}
\title{Access the vignettes}
\usage{
get_vignettes()
}
\description{
This function allows to access to the vignettes depsite them not being
installed properly. I had to do that as I want students to be able to install
the package quickly from github. They thus won't use a package bundle, nor I
want them to have to build the vignettes on their computers.
}
\examples{
get_vignettes()

}
