% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Gryphon}
\alias{Gryphon}
\title{The Gryphon dataset}
\format{
A list containing:
 \describe{
   \item{data}{The dataframe containing the mother id, the birth year, the sex, the birth weight, the tarsus length and the individual id of the gryphons}
   \item{pedigree}{The dataframe containing the individual id, the mother id and the father id of each gryphon}
 }
}
\source{
Alastair J. Wilson, Denis Réale, Michelle N. Clements, Michael M.
  Morrissey, Erik Postma, Craig A. Walling, Loeske E. B. Kruuk; Daniel H.
  Nussey; An ecologist's guide to the animal model. Journal of Animal
  Ecology 2010; 79, 13--26. doi: 10.1111/j.1365-2656.2009.01639.x
}
\usage{
Gryphon
}
\description{
This dataset contains a simulated dataset and a simulated pedigree of the
gryphons. The gryphon is a legendary creature with the body, tail, and back
legs of a lion; the head and wings of an eagle; and an eagle's talons as its
front feet. This dataset has been released with the publication of a paper
introducing the animal model, a particular LMM used to infer the heritability
of phenotypic traits (see sources). It thus offers the possibility to compare
different packages and software for estimating the heritability.
}
\keyword{datasets}
