% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Hedgehog}
\alias{Hedgehog}
\title{The Hedgehog dataset}
\format{
A data frame with 386 rows and 10 variables:
\describe{
  \item{time}{the date as a Date R object}
  \item{presence}{A binary indicating whether an hedgehog has been observed or not}
  \item{temperature}{the mean temperature during the 4-hour window}
  \item{humidity}{the mean humidity during the 4-hour window}
  \item{precipitation}{the precipitation amount over the 4-hour window?}
  \item{rain}{the proportion of hours with rain}
  \item{ground_temp}{the mean ground temperature during the 4-hour window}
  \item{cloud_cover}{the mean cloud cover during the 4-hour window}
  \item{moon_phase}{the phase of the moon on that night}
  \item{time_rel}{the number of days since the begining of the data collection}
}
}
\source{
Internship from Wanja Rast, IZW Berlin???
}
\usage{
Hedgehog
}
\description{
Data come from a camera trap that was set up in the same location for 2 years.
The column \var{presence} indicates whether an hedgehog was observed on this day or not.
All the other data correspond to measurements done in a 4-hour time frame
around sunset and all values are the mean over those 4 hours. The rain measurement
was a binary indicator for rain measured every hour so the values in this column
are the mean of this indicator. The cloud measurements were done categorically
on a scale from 0 to 8 so this is also the mean over 4 hours. Moon phase is also
categorical with 4 phases.
}
\keyword{datasets}
