% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HSE98women}
\alias{HSE98women}
\title{Sample from the Health Survey for England Cardiovascular Disease’98}
\format{
A data frame with 4 variables:
\describe{
  \item{period}{Whether still having periods}
  \item{age}{Age last birthday}
  \item{smoked}{Whether ever smoked cigarettes}
  \item{drunkLast3M}{Whether slightly (or very) drunk in the last three months}
  \item{drinkPb}{Number of drinking problems experienced}
  \item{education}{Age finished education}
  \item{bmi}{Valid BMI}
}
}
\source{
UK Data Archive Study Number 4150
}
\usage{
HSE98women
}
\description{
This dataset is a sample of few variables taken from the Health Survey for
England Cardiovascular Disease’98. This dataset is restricted to females.
}
\keyword{datasets}
