% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_Aliens}
\alias{simulate_Aliens}
\title{Function simulating the Alien datasets}
\usage{
simulate_Aliens(N = 12, intercept = 50, slope = 1.5, sigma2 = 25)
}
\arguments{
\item{N}{the number of Aliens (default = 12)}

\item{intercept}{the value for the intercep (default = 50)}

\item{slope}{the value for the slope (default = 1.5)}

\item{sigma2}{the value for the error variance (default = 25)}
}
\value{
a \var{data.frame} containing the simulated data
}
\description{
The Aliens considered here have a size that is varies linearly according to
the number of humans they have eaten.
}
\examples{
simulate_Aliens()

}
