% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_Aliens_GLM}
\alias{simulate_Aliens_GLM}
\title{Function simulating the Alien datasets for GLM}
\usage{
simulate_Aliens_GLM(N = 100)
}
\arguments{
\item{N}{the number of Aliens (default = 100)}
}
\value{
a \var{data.frame} containing the simulated data
}
\description{
This function creates different response variables to be used for GLM.
}
\examples{
simulate_Aliens_GLM()

}
