% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_predictors_1C1G}
\alias{simulate_predictors_1C1G}
\title{Function simulating the predictors}
\usage{
simulate_predictors_1C1G(n, k)
}
\arguments{
\item{n}{the number of datapoints to simulate per group}

\item{k}{the number of group to simulate}
}
\value{
a \var{data.frame} containing the simulated data
}
\description{
Function simulating the predictors
}
\examples{
simulate_predictors_1C1G(n = 3, k = 4)

}
