% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Surprise}
\alias{Surprise}
\title{The Surprise dataset}
\format{
A dataframe with 3 variables:
  \describe{
    \item{type}{the type of gift}
    \item{price}{the price coverted in US$}
    \item{hapiness}{the continuous scale used to record the children's emotional response}
  }
}
\source{
Justin Matejka, George Fitzmaurice, Toronto Ontario Canada. DOI: http://dx.doi.org/10.1145/3025453.3025912
}
\usage{
Surprise
}
\description{
This dataset is the result of a controlled study in which 852 parents from one
canadian school were divided into six groups of equal size and given gifts of
different types (1 type per group) and of different costs (between ca 15 to 100
dollars). After delivering the gifts to their children the parents reported
the perceived joy of their children on a continuous scale between 0 and 100.
}
\keyword{datasets}
