% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TitanicSurvival}
\alias{TitanicSurvival}
\title{Survival of Passengers on the Titanic (from package carData)}
\format{
A data frame with 1309 observations on the following 4 variables.
\describe{
  \item{survived}{\code{no} or \code{yes}.} \item{sex}{\code{female} or \code{male}.}
  \item{age}{in years (and for some children, fractions of a year); age is missing for
  263 of the passengers.}
  \item{passengerClass}{\code{1st}, \code{2nd}, or \code{3rd} class.}
 }
}
\source{
Data set \code{titanic3} from
\url{https://biostat.app.vumc.org/wiki/Main/DataSets}.
}
\usage{
TitanicSurvival
}
\description{
Information on the survival status, sex, age, and passenger class of 1309
passengers in the Titanic disaster of 1912.
}
\details{
This is part of a larger data set compiled by Thomas Cason.  Many additional
details are given in the sources cited below.
}
\references{
\url{http://www.encyclopedia-titanica.org/}

F. E. Harrell, Jr. (2001) \emph{Regression Modeling Strategies} New York:
Springer.
}
\keyword{datasets}
