% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Bats2}
\alias{Bats2}
\title{The second Bats dataset}
\format{A data frame with 82 rows and 14 variables:
\describe{
  \item{location}{the locations were different for every mistnet}
  \item{date}{the date as a Date R object}
  \item{Species}{the species name}
  \item{Sex}{the sex, male or female}
  \item{Age}{the age in age class: Juvenile or Adult}
  \item{X15N_fur}{isotopic measurement}
  \item{X13C_fur}{isotopic measurement}
  \item{X15N_membrane}{isotopic measurement}
  \item{X13C_membrane}{isotopic measurement}
  \item{X15N_liver}{isotopic measurement}
  \item{X13C_liver}{isotopic measurement}
  \item{X15N_stomach_content}{isotopic measurement}
  \item{X13C_stomach_content}{isotopic measurement}
  \item{time_rel}{the number of days since the begining of the data collection}
}}
\source{
Data collected by Cecilia Kruszynski, Center of Nuclear Energy in Agriculture, Piracicaba, São Paulo.
}
\usage{
Bats2
}
\description{
The dataset provide isotopic measurements from tissues sampled from different
bat species in different locations. Each tissue has a different turnover
which can show the difference in the diet through time. We don't provide
discrimination factors for each tissue.
}
\keyword{datasets}
